Move KafkaChannel e2e and Rekt tests to different dirs
