+++ .github/workflows/kind-e2e.yaml
@@ -27,6 +27,8 @@ jobs:
         test-suite:
           - ./test/e2e
           - ./test/e2e_new
+          - ./test/e2e_channel
+          - ./test/e2e_new_channel
 
         # Map between K8s and KinD versions.
         # This is attempting to make it a bit clearer what's being tested.
@@ -38,7 +40,7 @@ jobs:
           - k8s-version: v1.22.2
             kind-version: v0.11.1
             kind-image-sha: sha256:f638a08c1f68fe2a99e724ace6df233a546eaf6713019a0b310130a4f91ebe7f
-          - test-suite: ./test/e2e
+          - test-suite: ./test/e2e_channel
             extra-test-flags: -channels=messaging.knative.dev/v1beta1:KafkaChannel
 
     env:
