+++ test/e2e_channel/channel_single_event_test.go
@@ -17,7 +17,7 @@
  * limitations under the License.
  */
 
-package e2e
+package e2e_channel
 
 import (
 	"context"
