+++ test/e2e_new_channel/kafka_channel_upstream_test.go
@@ -17,7 +17,7 @@
  * limitations under the License.
  */
 
-package e2e_new
+package e2e_new_channel
 
 import (
 	"testing"
