+++ test/e2e_new_channel/main_test.go
@@ -1,5 +1,5 @@
-//go:build e2e || deletecm
-// +build e2e deletecm
+//go:build e2e
+// +build e2e
 
 /*
  * Copyright 2020 The Knative Authors
@@ -17,12 +17,14 @@
  * limitations under the License.
  */
 
-package e2e_new
+package e2e_new_channel
 
 import (
 	"flag"
 	"os"
 	"testing"
+
+	"knative.dev/eventing/test/rekt/resources/channel_impl"
 
 	// Uncomment the following line to load the gcp plugin (only required to authenticate against GKE clusters).
 	_ "k8s.io/client-go/plugin/pkg/client/auth/gcp"
@@ -30,10 +32,7 @@ import (
 	"knative.dev/pkg/injection"
 	_ "knative.dev/pkg/system/testing"
 
-	"knative.dev/eventing/test/rekt/resources/broker"
 	"knative.dev/reconciler-test/pkg/environment"
-
-	"knative.dev/eventing-kafka-broker/control-plane/pkg/reconciler/kafka"
 )
 
 // global is the singleton instance of GlobalEnvironment. It is used to parse
@@ -45,8 +44,9 @@ func init() {
 	// environment.InitFlags registers state and level filter flags.
 	environment.InitFlags(flag.CommandLine)
 
-	// Force the broker class to be configured properly
-	broker.EnvCfg.BrokerClass = kafka.BrokerClass
+	// CHANNEL_GROUP_KIND=KafkaChannel.messaging.knative.dev;CHANNEL_VERSION=v1beta1
+	channel_impl.EnvCfg.ChannelGK = "KafkaChannel.messaging.knative.dev"
+	channel_impl.EnvCfg.ChannelV = "v1beta1"
 }
 
 // TestMain is the first entry point for `go test`.
